feat #33 : 삭제 여부를 재확인하기 위해 window.confirm 사용

+++ src/components/posted-page/DeleteButton/index.jsx
@@ -15,7 +15,9 @@ function DeleteButton({ id, handleDeleteMessage, stopBubbling }) {
 
   const handleDeleteClick = (e) => {
     stopBubbling(e);
-    handleDeleteData(id);
+    if (window.confirm('정말 삭제하시겠어요?')) {
+      handleDeleteData(id);
+    }
   };
 
   return (
